店铺数据权限DataPermissionHandlerImpl；bootstrap-select店铺数据联动；经营概况表工资合计；jquery.validate升级；

+++ src/main/java/com/bootdo/modular/data/domain/DataShop.java
@@ -19,12 +19,12 @@ public class DataShop extends BaseEntity {
      *
      */
     @TableId(type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 店铺名称
      */
-    private Integer no;
+    private Long no;
 
     /**
      * 店铺名称
